style: google font import - Inter, Arimo
Import google fonts in global CSS. Set global CSS root font family to
Inter. Set Tailwind CSS default sans-serif to Arimo.

The page should use Inter everywhere that `sans-serif` is not specified.
The intention is to use Inter for the UI (for the right arrow ligature),
and manually set `sans-serif` in content containers.