refactor: portfolio object to match gray-matter style
See the Card component Prop description to see the new structure of the
portfolio object. Moves metadata fields under a data object, and moves the
description field to the content key. Content/description is wrapped in
ReactMarkdown for rendering.

The sample portfolio.json file is modified to depict the new schema.
is moved into the posts directory.